refactor: migrate CheckIn and Contribution types
