+++ docs/docs/auto-docs/shared-components/CheckIn/CheckInMocks/variables/checkInQueryMock.md
@@ -4,6 +4,6 @@
 
 # Variable: checkInQueryMock
 
-> `const` **checkInQueryMock**: (\{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `eventId`: `string`; \}; \}; `result`: \{ `data`: \{ `event`: \{ `id`: `string`; `recurrenceRule`: `any`; \}; \}; \}; \} \| \{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `eventId`: `string`; \}; \}; `result`: \{ `data`: [`InterfaceAttendeeQueryResponse`](../../../../types/CheckIn/interface/interfaces/InterfaceAttendeeQueryResponse.md); \}; \})[]
+> `const` **checkInQueryMock**: (\{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `eventId`: `string`; \}; \}; `result`: \{ `data`: \{ `event`: \{ `id`: `string`; `recurrenceRule`: `any`; \}; \}; \}; \} \| \{ `request`: \{ `query`: `DocumentNode`; `variables`: \{ `eventId`: `string`; \}; \}; `result`: \{ `data`: [`InterfaceAttendeeQueryResponse`](../../../../types/shared-components/CheckIn/interface/interfaces/InterfaceAttendeeQueryResponse.md); \}; \})[]
 
 Defined in: [src/shared-components/CheckIn/CheckInMocks.ts:38](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/CheckIn/CheckInMocks.ts#L38)
