+++ docs/docs/auto-docs/types/AdminPortal/Contribution/interface/interfaces/InterfaceContriStatsProps.md
@@ -1,0 +1,39 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceContriStatsProps
+
+Defined in: [src/types/AdminPortal/Contribution/interface.ts:1](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/Contribution/interface.ts#L1)
+
+## Properties
+
+### highestAmount
+
+> **highestAmount**: `string`
+
+Defined in: [src/types/AdminPortal/Contribution/interface.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/Contribution/interface.ts#L4)
+
+***
+
+### id
+
+> **id**: `string`
+
+Defined in: [src/types/AdminPortal/Contribution/interface.ts:2](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/Contribution/interface.ts#L2)
+
+***
+
+### recentAmount
+
+> **recentAmount**: `string`
+
+Defined in: [src/types/AdminPortal/Contribution/interface.ts:3](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/Contribution/interface.ts#L3)
+
+***
+
+### totalAmount
+
+> **totalAmount**: `string`
+
+Defined in: [src/types/AdminPortal/Contribution/interface.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/AdminPortal/Contribution/interface.ts#L5)
