--- docs/docs/auto-docs/types/CheckIn/interface/interfaces/InterfaceModalProp.md
@@ -1,47 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: InterfaceModalProp
-
-Defined in: [src/types/CheckIn/interface.ts:23](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L23)
-
-## Properties
-
-### eventId
-
-> **eventId**: `string`
-
-Defined in: [src/types/CheckIn/interface.ts:25](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L25)
-
-***
-
-### handleClose()
-
-> **handleClose**: () => `void`
-
-Defined in: [src/types/CheckIn/interface.ts:26](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L26)
-
-#### Returns
-
-`void`
-
-***
-
-### onCheckInUpdate()?
-
-> `optional` **onCheckInUpdate**: () => `void`
-
-Defined in: [src/types/CheckIn/interface.ts:27](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L27)
-
-#### Returns
-
-`void`
-
-***
-
-### show
-
-> **show**: `boolean`
-
-Defined in: [src/types/CheckIn/interface.ts:24](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L24)
