--- docs/docs/auto-docs/types/CheckIn/interface/interfaces/InterfaceTableData.md
@@ -1,31 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: InterfaceTableData
-
-Defined in: [src/types/CheckIn/interface.ts:42](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L42)
-
-## Properties
-
-### checkInData
-
-> **checkInData**: [`InterfaceTableCheckIn`](InterfaceTableCheckIn.md)
-
-Defined in: [src/types/CheckIn/interface.ts:45](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L45)
-
-***
-
-### id
-
-> **id**: `string`
-
-Defined in: [src/types/CheckIn/interface.ts:44](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L44)
-
-***
-
-### userName
-
-> **userName**: `string`
-
-Defined in: [src/types/CheckIn/interface.ts:43](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L43)
