--- docs/docs/auto-docs/types/CheckIn/interface/interfaces/InterfaceUser.md
@@ -1,31 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: InterfaceUser
-
-Defined in: [src/types/CheckIn/interface.ts:1](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L1)
-
-## Properties
-
-### emailAddress
-
-> **emailAddress**: `string`
-
-Defined in: [src/types/CheckIn/interface.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L4)
-
-***
-
-### id
-
-> **id**: `string`
-
-Defined in: [src/types/CheckIn/interface.ts:2](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L2)
-
-***
-
-### name
-
-> **name**: `string`
-
-Defined in: [src/types/CheckIn/interface.ts:3](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/interface.ts#L3)
