--- docs/docs/auto-docs/types/CheckIn/type/type-aliases/CheckIn.md
@@ -1,81 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Type Alias: CheckIn
-
-> **CheckIn** = `object`
-
-Defined in: [src/types/CheckIn/type.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L4)
-
-## Properties
-
-### \_id
-
-> **\_id**: `string`
-
-Defined in: [src/types/CheckIn/type.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L5)
-
-***
-
-### allotedRoom?
-
-> `optional` **allotedRoom**: `string`
-
-Defined in: [src/types/CheckIn/type.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L6)
-
-***
-
-### allotedSeat?
-
-> `optional` **allotedSeat**: `string`
-
-Defined in: [src/types/CheckIn/type.ts:7](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L7)
-
-***
-
-### createdAt
-
-> **createdAt**: `Date`
-
-Defined in: [src/types/CheckIn/type.ts:8](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L8)
-
-***
-
-### event
-
-> **event**: [`Event`](../../../Event/type/type-aliases/Event.md)
-
-Defined in: [src/types/CheckIn/type.ts:9](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L9)
-
-***
-
-### feedbackSubmitted
-
-> **feedbackSubmitted**: `boolean`
-
-Defined in: [src/types/CheckIn/type.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L10)
-
-***
-
-### time
-
-> **time**: `Date`
-
-Defined in: [src/types/CheckIn/type.ts:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L11)
-
-***
-
-### updatedAt
-
-> **updatedAt**: `Date`
-
-Defined in: [src/types/CheckIn/type.ts:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L12)
-
-***
-
-### user
-
-> **user**: [`User`](../../../shared-components/User/type/type-aliases/User.md)
-
-Defined in: [src/types/CheckIn/type.ts:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L13)
