--- docs/docs/auto-docs/types/CheckIn/type/type-aliases/CheckInInput.md
@@ -1,41 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Type Alias: CheckInInput
-
-> **CheckInInput** = `object`
-
-Defined in: [src/types/CheckIn/type.ts:16](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L16)
-
-## Properties
-
-### allotedRoom?
-
-> `optional` **allotedRoom**: `string`
-
-Defined in: [src/types/CheckIn/type.ts:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L17)
-
-***
-
-### allotedSeat?
-
-> `optional` **allotedSeat**: `string`
-
-Defined in: [src/types/CheckIn/type.ts:18](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L18)
-
-***
-
-### eventId
-
-> **eventId**: `string`
-
-Defined in: [src/types/CheckIn/type.ts:19](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L19)
-
-***
-
-### userId
-
-> **userId**: `string`
-
-Defined in: [src/types/CheckIn/type.ts:20](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/CheckIn/type.ts#L20)
