--- docs/docs/auto-docs/types/Contribution/interface/interfaces/InterfaceOrgContriCardsProps.md
@@ -1,63 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: InterfaceOrgContriCardsProps
-
-Defined in: [src/types/Contribution/interface.ts:8](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Contribution/interface.ts#L8)
-
-## Properties
-
-### contriAmount
-
-> **contriAmount**: `string`
-
-Defined in: [src/types/Contribution/interface.ts:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Contribution/interface.ts#L13)
-
-***
-
-### contriDate
-
-> **contriDate**: `string`
-
-Defined in: [src/types/Contribution/interface.ts:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Contribution/interface.ts#L12)
-
-***
-
-### contriTransactionId
-
-> **contriTransactionId**: `string`
-
-Defined in: [src/types/Contribution/interface.ts:14](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Contribution/interface.ts#L14)
-
-***
-
-### id
-
-> **id**: `string`
-
-Defined in: [src/types/Contribution/interface.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Contribution/interface.ts#L10)
-
-***
-
-### key
-
-> **key**: `string`
-
-Defined in: [src/types/Contribution/interface.ts:9](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Contribution/interface.ts#L9)
-
-***
-
-### userEmail
-
-> **userEmail**: `string`
-
-Defined in: [src/types/Contribution/interface.ts:15](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Contribution/interface.ts#L15)
-
-***
-
-### userName
-
-> **userName**: `string`
-
-Defined in: [src/types/Contribution/interface.ts:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Contribution/interface.ts#L11)
