+++ docs/docs/auto-docs/types/shared-components/CheckIn/interface/interfaces/InterfaceAttendeeCheckIn.md
@@ -1,0 +1,55 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceAttendeeCheckIn
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:7](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L7)
+
+## Properties
+
+### checkInTime
+
+> **checkInTime**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L10)
+
+***
+
+### checkOutTime
+
+> **checkOutTime**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L11)
+
+***
+
+### id
+
+> **id**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:8](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L8)
+
+***
+
+### isCheckedIn
+
+> **isCheckedIn**: `boolean`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L12)
+
+***
+
+### isCheckedOut
+
+> **isCheckedOut**: `boolean`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L13)
+
+***
+
+### user
+
+> **user**: [`InterfaceUser`](InterfaceUser.md)
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:9](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L9)
