+++ docs/docs/auto-docs/types/shared-components/CheckIn/interface/interfaces/InterfaceAttendeeQueryResponse.md
@@ -1,0 +1,23 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceAttendeeQueryResponse
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:16](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L16)
+
+## Properties
+
+### event
+
+> **event**: `object`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L17)
+
+#### attendeesCheckInStatus
+
+> **attendeesCheckInStatus**: [`InterfaceAttendeeCheckIn`](InterfaceAttendeeCheckIn.md)[]
+
+#### id
+
+> **id**: `string`
