+++ docs/docs/auto-docs/types/shared-components/CheckIn/interface/interfaces/InterfaceTableCheckIn.md
@@ -1,0 +1,79 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceTableCheckIn
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:30](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L30)
+
+## Properties
+
+### checkInTime
+
+> **checkInTime**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:34](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L34)
+
+***
+
+### checkOutTime
+
+> **checkOutTime**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:35](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L35)
+
+***
+
+### eventId
+
+> **eventId**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:38](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L38)
+
+***
+
+### id
+
+> **id**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:31](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L31)
+
+***
+
+### isCheckedIn
+
+> **isCheckedIn**: `boolean`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:36](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L36)
+
+***
+
+### isCheckedOut
+
+> **isCheckedOut**: `boolean`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:37](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L37)
+
+***
+
+### isRecurring?
+
+> `optional` **isRecurring**: `boolean`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:39](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L39)
+
+***
+
+### name
+
+> **name**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:32](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L32)
+
+***
+
+### userId
+
+> **userId**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:33](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L33)
