+++ docs/docs/auto-docs/types/shared-components/CheckIn/interface/interfaces/InterfaceTableData.md
@@ -1,0 +1,31 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceTableData
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:42](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L42)
+
+## Properties
+
+### checkInData
+
+> **checkInData**: [`InterfaceTableCheckIn`](InterfaceTableCheckIn.md)
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:45](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L45)
+
+***
+
+### id
+
+> **id**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:44](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L44)
+
+***
+
+### userName
+
+> **userName**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:43](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L43)
