+++ docs/docs/auto-docs/types/shared-components/CheckIn/interface/interfaces/InterfaceUser.md
@@ -1,0 +1,31 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceUser
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:1](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L1)
+
+## Properties
+
+### emailAddress
+
+> **emailAddress**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L4)
+
+***
+
+### id
+
+> **id**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:2](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L2)
+
+***
+
+### name
+
+> **name**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/interface.ts:3](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/interface.ts#L3)
