+++ docs/docs/auto-docs/types/shared-components/CheckIn/type/type-aliases/CheckInStatus.md
@@ -1,0 +1,33 @@
+[Admin Docs](/)
+
+***
+
+# Type Alias: CheckInStatus
+
+> **CheckInStatus** = `object`
+
+Defined in: [src/types/shared-components/CheckIn/type.ts:23](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/type.ts#L23)
+
+## Properties
+
+### \_id
+
+> **\_id**: `string`
+
+Defined in: [src/types/shared-components/CheckIn/type.ts:24](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/type.ts#L24)
+
+***
+
+### checkIn?
+
+> `optional` **checkIn**: [`CheckIn`](CheckIn.md)
+
+Defined in: [src/types/shared-components/CheckIn/type.ts:25](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/type.ts#L25)
+
+***
+
+### user
+
+> **user**: [`User`](../../../User/type/type-aliases/User.md)
+
+Defined in: [src/types/shared-components/CheckIn/type.ts:26](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/CheckIn/type.ts#L26)
