+++ src/types/shared-components/CheckIn/interface.ts
